Change to int

+++ src/main/java/com/czertainly/core/dao/repository/notifications/NotificationProfileVersionRepository.java
@@ -48,6 +48,6 @@ public interface NotificationProfileVersionRepository extends SecurityFilterRepo
             WHERE npv.notificationInstanceRefUuid = :uuid
             """)
     @Modifying
-    Long detachNotificationInstanceRefUuid(@Param("uuid") UUID uuid);
+    int detachNotificationInstanceRefUuid(@Param("uuid") UUID uuid);
 }
